Update DAL for vespa bugfixes (#189)
* Update fastapi & starlette

In order to update the data acess library, we also need to update fastapi
This also means updating fastapi (and starlette) to avoid clashing pydantic
versions. FastAPI releases: https://github.com/tiangolo/fastapi/releases

* Bump poetry to handle build error

This is in response to the following error:
```
ERROR: pip's dependency resolver does not currently take into account
all the packages that are installed. This behaviour is the source of the
following dependency conflicts.
poetry 1.2.2 requires platformdirs<3.0.0,>=2.5.2, but you have
platformdirs 4.0.0 which is incompatible.
```

* DAL for vespa bugfixes

DAL releases: https://github.com/climatepolicyradar/data-access/releases

* Test downgrading DAL version

* Update to latest DAL version

* Include extras from DAL

* Simplify export command

Poetry now has a handy plugin for exporting to a requirements.txt, this
means we can move away from our own custom version

* Explicitly install sentence transformers

This dockerfile downloads a model from sentence transformers, previously
we installed sentence transformers implicitly as part of the install of
the data access library (DAL), in more recent versions of the DAL, we
moved this dependency under the vespa extra. This now becomes a problem
for the export command, which says: "this command only includes the
project's dependencies defined in the implicit main group"[^1].

[^1]: [Export Command](https://python-poetry.org/docs/cli/#export)

* Updating pydantic.

* Swith to using optional extras in DAL

As groups clashed with the way we install dependencies in the backend

* Update failing taxonomy validation

* Updata extra handling for deprecation warnings

Because of this warning:
```
pydantic.config.Extra is deprecated, use literal values instead (e.g.
extra='allow'). Deprecated in Pydantic V2.0 to be removed in V3.0.
```

* Updating dependency reference.

* Updating the dal tag.

---------

Co-authored-by: Mark <mark@climatepolicyradar.org>