Switch searches to use vespa by default (#205)
* Switch searches to use vespa by default

Minimal change switchover to make it easy to switch back if needed, this
just makes vespa the default over opensearch. The api can still query
opensearch by setting `use_vespa` to False. We can delete the opensearch
code when we're confident we won't need it as a fallback

* Keep opensearch tests working

This distinguishes opensearch tests by giving them their own pytest mark,
it also adds a parameter to the api request to cause the endpoint to still
use opensearch for these tests. Finally we rename the make task for
opensearch specific tests

* Setup for vespa in CI

Adds test db setup, fixtures and commands to run vespa in ci

* Direct most opensearch tests towards vespa

Some have been made obsolete, either overtime or because vespa handles
something differently. But this adapts as many of the existing opensearch
tests as possible to point at the vespa route instead

* Tidy up make file commands

* remove redundant test

No longer needed as we store the slug in vespa

* increase disk space folllowing feed errors in github actions

* Remove filter keyword tests until issue resolved

These tests are currently failing due to an issue in the Data access
library. Removing them until the issue can be resolved

* Re add filter tests following DAL fix

Context: https://github.com/climatepolicyradar/data-access/pull/118

* remove test override used to confirm dal changes

* add type hints to vespa test setup

* Move search test setup to their own file

* Tidy up makefile variable