Reluctantly move ingest code into tests for setup
Unfortunatly we rely on the ingest code for test setup. So although this
has been removed from the app, we still need it to maintain test coverage.
That being said, we are one step closer to deleting it and no longer need
to test it, so this is still progress.

The ideal solution here probably to replace the many fragmented test
setup with a single, repurposable setup factory. But thats beyond the
scope of the current change