Feature/pdct 774 remove csvs from backends pipeline trigger (#216)
* Remove reduntant test code

We no longer use the csv ingest process so don't need to waste test time
on it

* Remove unfccc ingest route

* Refacter cclw ingest to be an ingest trigger

This had a bunch of overhang from before the admin service. Turning it
into a dedicated trigger makes it easier to call and run

* Remove as much legacy ingest code as we can

Some of these paths are still used by other code. Leaving them in for now
is the path of least resistance for getting this done and gets closer to
a tidier state. With any luck we'll be able to take it further next time

* Reluctantly move ingest code into tests for setup

Unfortunatly we rely on the ingest code for test setup. So although this
has been removed from the app, we still need it to maintain test coverage.
That being said, we are one step closer to deleting it and no longer need
to test it, so this is still progress.

The ideal solution here probably to replace the many fragmented test
setup with a single, repurposable setup factory. But thats beyond the
scope of the current change

* Clean up redundant mocks

These are no longer needed following the previous commit

* Rename `cclw_ingest` to `pipeline_trigger`