Automatically run the pipeline on a weekly schedule
Triggers the db state file creation endpoint every monday at 9:15. This
means we can run during work hours for the time being as we grow in
confidence in the pipeline run, later it can be moved to run overnight
instead.

The implementation is intended to be simple and could be moved to a
seperate service from the backend in the future.