Make filters exclusive to each other
This surfaces the changes from the data access library:
https://github.com/climatepolicyradar/data-access/pull/141

As part of that it also fixes an issue with country & region searches,
where country codes from both would be combined into one, so we'd always
get the whole region, rather refining within it.

Note that there was two functions handling this logic, one for search and
another for browse (which had no tests), I've partially consolidated them
but had to add a workaround until we come back and do this properly.