Move cron job to startup
We had the problem where three of these jobs started at the same time,
likely because there where three backend instances running at the time
due to the backend being at its scaling limit. The thinking is when the
parent process forks it is also brings the cron job with it.

The hope here is that moving it to the startup will cause it to only run
once. As an extra check this also adds an id and sets the job to replace
any existing jobs with that id.

It's hard to test this to know if it will work for sure, but it is
already broken, and if it turns out this still doesnt work I'll just
switch back to triggering it manually until there is enough time to do a
better solution