Simplify request params by extending DAL version
Loading one pydantic model into another is inefficient. This changes
the search request parameters to (as much as it can) use the data access
library's version instead.

The exception is where behaviour diverges. Filters need to be converted
from region/country names to iso codes, page size and offset still need
to be supported for the time being.