Add initial documentation for search endpoint
This is limited ot the search endpoint for now, as it is the most used.
At the time of this commit it is not possible to make the `search_in_schema`
default to false. But this might be a neater way of doing it in the future,
see: https://github.com/tiangolo/fastapi/discussions/9226

This work also clashes with some deprioritised work to move these models
to the sdk. That work is intended to simplify the code and make the
experience for data scientists more consistent with search users. However,
other urgent work meant it was left with some fields defined in both. See:
https://linear.app/climate-policy-radar/issue/PDCT-1024/parameter-handling-needs-deduplicating-response