Detect ingest_cycle_start automatically
This removes a manual task to update the ingest date after every ingest.
Instead we deduce the last ingest from the pipeline cache. The backend
already interacts with the pipeline cache, so this ends up meaning
extending the functionality to search through the bucket for when a
db_state.json file was most recently written.