Simplify dockerfile
The previous version of the dockerfile saugh tto solve a size problem
caused by torch. Torch is an implicit dependency from the sdk and is
required for embeddings. The GPU version of torch is huge. And would break
app runner if we tried to include it, we only need the cpu version so we
hard coded the installation for that in the dockerfile.

Aside from looking a little messy, this made it difficult to install
dependencies outside of docker, and managing depdendencies this way
leaves extra room for versioning issues.

This moves the torch dependency to poetry so it can be more explicitly,
and more consistently, managed. Note that this requires accounting for
mac specific torch install behaviour.