Move session to function scope
Longer running test runs where routinely seeing the error:
```
sqlalchemy.exc.ResourceClosedError: This Connection is closed
```

We previously had the connection shared across the whole test run, which
is faster than starting a new connection each time, unfortunatly it
seems to be leading to an issue with the session closing prematurely.
To fix the issue, this moves the connection to be created on a per test
function basis.

The tradeoff is this slows a full test run with `make test` down slightly,
but this seems better than them not working.