Alias `limit` to `page_size` so we can swap them
We want to change the api to use `page_size` instead of `limit`, so that
limit can be adjusted and can be used consistently across the backend and
vespa. Step one of doing this is aliasing `limit` with `page_size` so
we can change the frontend over and then return to change the backend.
This means we can make the change without downtime.