Fix retrieving document type from metadata
The pipeline trigger was failing due to a validation error on the document
type:

```
1 validation error for DocumentParserInput
type
  Input should be a valid string [type=string_type, input_value=None, input_type=NoneType]
```

There was an update last week to how we determine the document type. It
seems there are situations when this can evaluate to None. Here we
extract the logic into its own function so we can apply a fix and test
various conditions.