Use sdk model for pipeline input
The two models are identical and at risk of diverging. This makes it so
we can consolidate where updates need to happen to just the sdk model.

Note that there was the small config difference already which I have
preserved by inheriting the sdk model into the existing backend model.